Move tests to core-bpf-migration module
