+++ runtime/src/bank.rs
@@ -200,7 +200,6 @@ struct VerifyAccountsHashConfig {
 mod accounts_lt_hash;
 mod address_lookup_table;
 pub mod bank_hash_details;
-mod builtin_programs;
 pub mod builtins;
 mod check_transactions;
 mod fee_distribution;
